+++ CHANGELOG.md
@@ -14,6 +14,13 @@ NOTE: This is the CHANGELOG for the @looker/components MONOREPO. Each package ha
 - [Design Tokens](./packages/design-tokens/CHANGELOG.md)
 - [Icons](./packages/icons/CHANGELOG.md)
 
+
+## [UNRELEASED]
+
+### Added
+
+- DataTable PrimaryAction support
+
 ## [0.9.25]
 
 ### Changed